I realized that the transformations may need to modify more than each data row in isolation.  For example, in adding a "grade" column there is also the need to add a header for that new column.
So the transformation is now not simply a function but an
associative array of functions where the key of each mapping
is the row served by the associated function.  For example,
data rows are handled by the function mapped from "dataRow"
while the header row is handled by the function mapped
from "headerRow".

This is an associative array instead of a list to make it
easier to add later additional types of transforming functions
along with transforms that might have only a subset of those
functions.  For example, a transform which merely changes data
within a column need not have a headerRow function.