Begin a file which will apply conversions to the student data file.
We're now receiving data from Genesis instead of Skyward.  More,
a number of fields we've been requesting for a while have been added.
However, it appears that it may not be possible for us to get the
PTA software upgraded to match on a timely basis.  I'm therefore
starting a program which will convert the better data from Genesis
into the unfortunate format we used to receive from Skyward.

This is just the beginning, which is starting with a copy of the
file which used to convert the Skyward staff data file.