The start of a program to load student data into a SQL database.
I'm aware of the inefficiency of the fact that the source of this
data - Genesis - and the destination of this data both use SQL-based
RDBs, with the intermediate data encapsulation being a spreadsheet.
The district is running a program that denormalizes the data, wiping away
most structure and storing it as a flat spreadsheet.  Then there is
this program to read that flat spreadsheet and resurrect the implied
structure.

At some point, I hope to negotiate with the district a more well-structured
data exchange format.

For now, though, this is what we have as this is what the programmer
of the original site (or perhaps one of the PTA volunteers) understood
to request.

This is just the beginning of this program.  It is essentially
convertStudentFileFromGenesis.pl with all the "spreadsheet output"
logic removed. If any of this actually goes into production, the
shared parsing logic should be moved into a library.