This is writing data to school and school_name tables.
At the moment, this is working "row-first".  That is, it
processes each student's record/row in turn.  There's no
"pre-filtering" of school information in local memory;
all duplicate checking is in the server, for example.

This is the simplest approach, but almost certainly not the
most efficient.  Later, I may want to do more local
processing and send only the results to the server.  On the
other hand, that approach introduces race conditions.
At a minimum, some form of optimistic concurrency would
need to be exploited (locking would be a poor choice as
the race condition will hopefully be infrequent and the
locks would have to be held for a long time).

I may be able to avoid the race conditions.

A function is called to try to add the school's information
to the DB.  It succeeds whether the INSERT occurs or isn't
necessary as the school's data is there.

Either way, the DB's ID for the school is returned.  This
will support subsequent insertion of additional data for
the row being processed.

This is successfully using mysql_multi_statements.

+       # Trying mysql_multi_statements.
+       # Would mysql_server_prepare improve performance?
+       # Also consider processing in "column first" order.  That is: all school data, then all student data, then all con
+       # This permits multivalued inserts if one assumes that this is loading into a clean/empty database.