Add simple transaction support, and avoid "fetch() without execute()" errors.
The transaction support is per-row.  Autocommit is disabled, and the
transaction is committed (or rolled back) when the row's processing
is complete.

Avoiding the "fetch() without execute()" involved checking a result
set's column count as well as the row count.  It turns out that a
result set involving an insert returns a row count of 1 (or perhaps
more for multivalued inserts).  Retrieving this row fails, though,
as the number of columns is zero.