The student table is being populated.
I want to give serious thought to shifting from "insert ignore"
to "insert ... on duplicate update...".  This would permit
an insert/update which leaves the identifying values alone
but which permits update of (for example) grade or last name
or some such, in the case of the student table.

This is nonstandard SQL, but there's an equivalent feature
in PostgreSQL (at least as of v9.5): "UPSERT" which is "insert ...
on conflict do upate ... ".

9.5 is relatively recent, though, so hosting vendors may not
yet have it.  I seem to recall Heroku, for example, is on 9.4
at the moment.

https://wiki.postgresql.org/wiki/UPSERT