Add population of homeroom table.
Student now joins not to school but to homeroom, and homeroom joins
to school.  This seems reasonable for the vast majority of the cases.
However, I noticed that the schema used by Genesis

http://www.plainfieldnjk12.org/Departments/Genesis/DATA%20DICTIONARY/Genesis%20Data%20Dictionary%20version%202.pdf

does not work that way.  There, student joins to school and homeroom
individually.  It is possible for a student to be "in" one school but
have a homeroom in another.

Perhaps we should support that model, Just In Case?

This commit also includes handling of a student that has blanks for
homeroom and teacher.  In the subset of the data I've been using for
testing, I found this in student 99907457.