Student joins to both homeroom and school.
I discussed in ccee5cdfb5db413d7746d471f97af5bb47af690e the fact that
the Genesis schema works this way.  Given that, it is not unreasonable
for our schema to support this as well.

Note that the current mechanism for data transfer from Genesis to
us would *not* permit this (technically, it would break), but that
spreadsheet is subject to change.