The district has changed how it sends student data.  The "main" student file is now only to be used for email broadcasts.  It therefore no longer contains columns associated with physical addresses.
Happily - because it uses column names rather than locations - this
does not impact loadStudentFileIntoSQL.pl.

A less good reason - because loadStudentFileIntoSQL.pl doesn't yet
handle address - is also a reason.

We need a separate program to handle a new file which contains
physical addresses and is to be used for the purpose of building
printed directories.

I believe that we'll need to decompose student_contact into at
least two tables, one of which contains email information and
the other physical address.  There may be common data to be
stored in a third table to which the first two join.