As discussed in #2 and #3, student_contact is severely decomposed. This separates email, phone, and address information and makes those entities sharable amongst multiple contacts.
Note that this avoids the "update anomaly" normally mentioned in
normalization discussions, but it does create a new version of the
anomaly where two people that previously shared a phone or address
cease doing so.  We'll learn of this by a change of phone or address
but we'll likely have no way to discern that this requires a
separation of records rather than a change to the shared record.

I did make use of boolean columns directly in tables, such as
"use_in_directory" and "use_in_broadcast" in student_contact, rather
than separate attribute tables.  I consider this technical debt that
I may come to regret, but it'll let me move forward a bit more
quickly to something usable.