Added population of phone rows and joining to student_contact.  This required adding additional uniqueness constraints.
Note that "primary phone" uses the column/flag named "prime" as
"primary" is a keyword.

Also note that I'm not entirely sure that the model is correct.  It
certainly does not permit expression of exactly what the spreadsheets
express.  In the spreadsheet, the properties of "primary phone",
"cell phone", and "home phone" are per-contact.  In this DB, these are
stored per-phone number.

I have found instances where different contacts use the same phone
numbers but with different properties.