The test program is running a sample query and writing this to a sheet.
The query is just something I was using to examine the data; it is
of no consequence.

But the results of the query are being written pretty well.  There's no
format control of the sheet yet (no alignment, column size, etc.) but
the column headers are being written as one row of the sheet (the first
row of the sheet for the first result set from the query - and for now,
there is only one result set in the query), and the data rows are written
in subsequent rows of the sheet.

This is exactly the sort of thing we'll want.

This commit added the usual "db-centric" options to the program,
permitting specification of the db host, username, port, etc.