Updated the query to handle both "place in directory" and "use for contact" uses.
First, the select list is cleaner about null values.

If email or phone information is missing, this doesn't prevent the record
from appearing thanks to outer joins.

The query takes input indicating whether a record should be checked
for its "want to be in directory" or "wants to be used for contact"
property (this may be better done a different way).