Avoid using internal school ID in command line; use district's ID instead. *Never* use internal (semantic-free) IDs externally.  Aside from anything else, they are not stable over time in our case (as the DB is reloaded frequently).
Note that this required changing the command line option from a numeric
value to a string since the ID of the DLC is "DLC" (which is, as best
I can determine, not a number unless we're speaking of base 22).

Also adding convenience scripts to generate school-specific
files for all schools.