This is a variation of outputContactsAsXLXS.pl which outputs CSV and also outputs grades 8-11 for all schools.  It is intended as a work-around to let us reach HS students before the district has provided us with this year's data.  It therefore uses last year's data, but assumes that everyone has been promoted and that ex-8th-graders have all moved into the high school.
It also outputs CSV instead of XLXS, which means that the name
should probably be changed.  This is because we're no longer
populating the old web site but instead something more modern
(ie. mailchimp or memberhub).

The XLXS code remains at least for now, though, in case we need that
other format.