Fix WHERE of student_contact_phone rows, permitting contacts missing one or more of the phone numbers.  This required moving conditions checking the home, cellular, and prime columns to the ON clause.  Placing those conditions in the WHERE clause effectively removed the "outerness" of the joins because a NULL in student_contact_phone.cell (for example) would cause a row to mismatch.
This problem arose because student_contact_phone.cell has NULL rather
than zero on a phone that is not a cell phone.  That's not really correct,
as it turns out, and it blocks the WHERE clause from differentiating
between "no data" and "not a cell phone".

So, while this fix works, a perhaps better approach would be setting
the columes to zero instead of null when a phone is not of that type.

This solves a problem raised in 85f72fa98befacdf90a55e26c96bcfbdf15d7d9f,
where Amy's contact information was missing.  She has no "home" number
in the database.